Adapt to Bioimage zoo models format. Robustness
- Almost complete adoption of the Bioimage model zoo requirements
- Version control of models
- Small changes in interface
- Correction of small bugs

+++ src/main/java/deepimagej/stamp/SaveStamp.java
@@ -42,6 +42,7 @@ import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.Font;
 import java.awt.Frame;
+import java.awt.TextField;
 import java.awt.datatransfer.DataFlavor;
 import java.awt.datatransfer.Transferable;
 import java.awt.datatransfer.UnsupportedFlavorException;
@@ -52,8 +53,12 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Set;
+import java.util.Vector;
 
 import javax.swing.BorderFactory;
 import javax.swing.JButton;
@@ -68,10 +73,12 @@ import deepimagej.DeepImageJ;
 import deepimagej.Parameters;
 import deepimagej.components.HTMLPane;
 import deepimagej.tools.FileTools;
+import deepimagej.tools.Index;
 import deepimagej.tools.YAMLUtils;
 import ij.IJ;
 import ij.ImagePlus;
 import ij.WindowManager;
+import ij.gui.GenericDialog;
 import ij.measure.ResultsTable;
 import ij.text.TextWindow;
 
@@ -161,11 +168,46 @@ public class SaveStamp extends AbstractStamp implements ActionListener, Runnable
 		params.saveDir = params.saveDir.replace(File.separator + File.separator, File.separator);
 		File dir = new File(params.saveDir);
 		boolean ok = true;
+		ArrayList<String> repeatedFiles = new ArrayList<String>();
 		if (!dir.exists()) {
 			dir.mkdir();
 			pane.append("p", "Make a directory: " + params.saveDir);
+		} else {
+			repeatedFiles = checkFolder(params.saveDir);
+		}
+		
+		if (repeatedFiles.size() > 0 && repeatedFiles.size() < 3) {
+			GenericDialog dlg = new GenericDialog("Overwrite existing model");
+			dlg.addMessage("Part of the model was found in the save directory.");
+			dlg.addMessage("Do you want to overwrite it?.");
+			dlg.addMessage("If you press Ok the following files/folders will be deleted/overwritten.");
+			if (new File(dir, "variables").isDirectory())
+				repeatedFiles.add("variables");
+			for (String f : repeatedFiles)
+				dlg.addMessage("- " + f);
+			dlg.showDialog();
+			if (dlg.wasCanceled()) {
+				pane.append("p", "Save cancelled");
+				return;
+			} 
+			for (String f : repeatedFiles) {
+				boolean deleted = true;
+				if (f.contains(".")) {
+					deleted = new File(f).delete();
+				} else {
+					deleted = FileTools.deleteDir(new File(f));
+				}
+				if (!deleted) {
+					IJ.error("Could not remove " + f + ".\nSave cancelled.");
+					pane.append("p", "Save cancelled");
+				}
+			}
 		}
 		dir = new File(params.saveDir);
+		
+		// Check if save directory already has a version of the same model. If it does
+		// keep both models together and indicate it in the YAML file
+		
 		if (ok && !dir.exists()) {
 			IJ.error("This directory is not valid to save");
 			ok = false;
@@ -177,12 +219,83 @@ public class SaveStamp extends AbstractStamp implements ActionListener, Runnable
 
 		// Save the model architecture
 		try {
-			File source = new File(params.path2Model + "saved_model.pb");
-			File dest = new File(params.saveDir  + "saved_model.pb");
-			FileTools.copyFile(source, dest);
+			// Check that, if the save directory is already a Bioimage Model Zoo model.
+			// If it is check that they have the same architecture
+			boolean overwrite = true;
+			if (params.biozoo) {
+				int sameModel = isItTheSameModel(params.saved_modelSha256, params.saveDir + File.separator + "saved_model.pb", params.path2Model + "saved_model.pb");
+				if (sameModel == 1) {
+					// Do not save anything as it is the same
+					overwrite = false;
+				} else if (sameModel == 0) {
+					GenericDialog dlg = new GenericDialog("Overwrite existing model");
+					// TODO change to official name
+					// TODO check if we created two different models at different times
+					// TODO with the same archi, the sha256 will be the same
+					dlg.addMessage("There is already a Biozoo model in the save directory.");
+					dlg.addMessage("Do you want to store the current model as another version of the model already saved?.");
+					dlg.addMessage("In this case the architecture from the already saved model will be kept for both the old and new weights.");
+					dlg.addMessage("You can also overwrite the files with the new model?.");
+					dlg.addChoice("Select", new String[] {"Overwrite",  "Same version"}, "Overwrite");
+					dlg.showDialog();
+					if (dlg.wasCanceled()) {
+						pane.append("p", "Save cancelled");
+						return;
+					} 
+					String choice = dlg.getNextChoice();
+					if (choice.contains("Same version"))
+						overwrite = false;
+				} else if (sameModel == -1) {
+					GenericDialog dlg = new GenericDialog("Overwrite existing model");
+					dlg.addMessage("There is already a Biozoo model in the save directory.");
+					dlg.addMessage("However the model seems to have been modified and it cannot longer be loaded.");
+					dlg.addMessage("Do you want to overwrite it?.");
+					dlg.showDialog();
+					if (dlg.wasCanceled()) {
+						pane.append("p", "Save cancelled");
+						return;
+					} 
+					dlg = new GenericDialog("Overwrite existing weights");
+					dlg.addMessage("What do you want to do with the weights?");
+					dlg.addMessage("If they work with the architecture you are saving currently,");
+					dlg.addMessage("you could keep them as another version of the model.");
+					dlg.addMessage("If they correspond to another architecture, it is better to delete them.");
+					dlg.addMessage("Do you want to delete them?");
+					dlg.enableYesNoCancel();
+					dlg.showDialog();
+					if (dlg.wasCanceled()) {
+						pane.append("p", "Save cancelled");
+						return;
+					} else if (dlg.wasOKed()) {
+						for (File w : dir.listFiles()) {
+							if (w.getName().contains("weights_v") && w.getName().contains(".zip")) {
+								boolean deleted = w.delete();
+								if (!deleted) {
+									IJ.error("Could not remove " + w.getName() + ".\nSave cancelled.");
+									pane.append("p", "Save cancelled");
+								}
+							}
+						}
+					} 
+					
+				} else if (sameModel == -2) {
+					overwrite = true;
+				} else if (sameModel == -3) {
+					overwrite = true;
+				}
+			}
+			if (overwrite) {
+				File source = new File(params.path2Model + "saved_model.pb");
+				File dest = new File(params.saveDir  + "saved_model.pb");
+				FileTools.copyFile(source, dest);
+			}
 			pane.append("p", "protobuf of the model (saved_model.pb): saved");
-		}
-		catch (Exception e) {
+		} catch (IOException e) {
+			e.printStackTrace();
+			pane.append("p", "protobuf of the model (saved_model.pb): not saved");
+			pane.append("p", "protobuf of the model (saved_model.pb): saved_model.pb was removed from the model path");
+			ok = false;
+		} catch (Exception e) {
 			e.printStackTrace();
 			pane.append("p", "protobuf of the model (saved_model.pb): not saved");
 			ok = false;
@@ -190,10 +303,35 @@ public class SaveStamp extends AbstractStamp implements ActionListener, Runnable
 
 		// Save the model weights
 		try {
+			if (params.biozoo) {
+				// Check that the version string is not already in the yaml, if there is
+				//  a yaml in the directory.
+				boolean repeated = getModelVersion(params.saveDir, params);
+				while (repeated) {
+					GenericDialog dlg = new GenericDialog("Change version of the weights");
+					dlg.addMessage("Version " + params.version + " already exists. Choose a different one.");
+					dlg.addStringField("New weights version", "X.Y");
+					dlg.showDialog();
+					if (dlg.wasCanceled()) {
+						pane.append("p", "Save cancelled");
+						return;
+					}
+					Vector<TextField> strField = dlg.getStringFields();
+					TextField versionField = (TextField) strField.get(0);
+					params.version = versionField.getText().trim();
+					repeated = getModelVersion(params.saveDir, params);
+					if (repeated) {
+						IJ.error("This version already exists");
+					} 
+					
+				}
+			}
+			String zipName = "weights_v" + params.version + ".zip";
 			File source = new File(params.path2Model + "variables");
-			File dest = new File(params.saveDir + "variables");
-			copyWeights(source, dest);
+			File dest = new File(params.saveDir + zipName);
+			FileTools.zipFolder(source, dest);
 			pane.append("p", "weights of the network (variables): saved");
+			// TODO add check to ensure each pair of weights is unique
 		}
 		catch (Exception e) {
 			e.printStackTrace();
@@ -297,20 +435,77 @@ public class SaveStamp extends AbstractStamp implements ActionListener, Runnable
 
 		//parent.setEnabledBackNext(ok);
 	}
-
-	private void copyWeights(File source, File dest) throws IOException {
-		String source_path;
-		String dest_path;
-		String filename;
-		File[] n_files = source.listFiles();
-		for (int i = 0; i < n_files.length; i++) {
-			if (n_files[i].isFile() == true) {
-				filename = n_files[i].getName();
-				source_path = source.getAbsolutePath() + File.separator + filename;
-				dest_path = dest.getAbsolutePath() + File.separator + filename;
-				FileTools.copyFile(new File(source_path), new File(dest_path));
+	
+	/*
+	 * Check that the architecture in the old model is correct per the yaml, and if
+	 * it is the same as the new architecture we want to save
+	 * Returns 1 if the model specified in he yaml is the same as the one we want to save
+	 * and the one that is previously in the dir; 0 if the one specified in the yaml
+	 * is the same as the one already in the folder, but not the same as the new one we
+	 * want to save; -1 if none of them is the same; -2 if the one specified by the yaml
+	 * is the same as the one we want to save now but not the same as the one that is already
+	 * and -3 if the old model is missing from the directory.
+	 */
+	public static int isItTheSameModel(String originalSha256, String oldArchiDir, String newArchiDir) throws IOException{
+		String oldSha256 = "";
+		String newSha256 = FileTools.createSHA256(newArchiDir);
+		if (new File(oldArchiDir).isFile()) {
+			oldSha256 = FileTools.createSHA256(oldArchiDir);
+		} else {
+			return -3;
+		}
+		if (originalSha256.equals(oldSha256) && originalSha256.equals(newSha256)) {
+			return 1;
+		} else if (originalSha256.equals(oldSha256) && !originalSha256.equals(newSha256)) {
+			return 0;
+		} else if (!originalSha256.equals(oldSha256) && !originalSha256.equals(newSha256)) {
+			return -1;
+		} else {
+			// !originalSha256.equals(oldSha256) && originalSha256.equals(newSha256)
+			return -2;
+		}
+	}
+	
+	/*
+	 * Return all the versions of the model saved in the folder.
+	 * The method looks at the yaml, checks the versions that are already saved
+	 * and makes sure they are correct with the sha256 
+	 */
+	public static boolean getModelVersion(String dir, Parameters params){
+		
+		//Map<String, Object> obj =  YAMLUtils.readConfig(dir + File.separator + "config.yaml");
+		Set<String> keyset = params.previousVersions.keySet();
+		String[] versions = new String[keyset.size()];
+		int c = 0;
+		// Each version name consists of 'v' + 'version string'
+		boolean repeated = false;
+		for (String key : keyset) {
+			versions[c ++] = key.substring(1);
+			if (key.substring(1).equals(params.version.trim())) {
+				repeated = true;
 			}
 		}
+		return repeated;
+	}
+	
+	public static ArrayList<String> checkFolder(String dir){
+		ArrayList<String> files = new ArrayList<String>();
+		if (new File(dir, "saved_model.pb").isFile())
+			files.add("saved_model.pb");
+		if (new File(dir, "config.yaml").isFile()) {
+			files.add("config.yaml");
+		} else if (new File(dir, "config.yml").isFile()) {
+			files.add("config.yml");
+		}
+		String[] fList = new File(dir).list();
+		for (String f : fList) {
+			if (f.contains("weights_v") && f.contains(".zip")) {
+				files.add(f);
+				break;
+			}
+		}
+		
+		return files;
 	}
 
 	public class LocalDropTarget extends DropTarget {
